refactor: App.js에서 Console 영역을 ConsoleApp.js로 분리
